+++ NEWS.md
@@ -1,4 +1,4 @@
-# rolap (development version)
+# rolap 2.5.0
 * Geographic information query functions: `as_geolayer()`, `get_layer()`, `get_variables()`, `set_variables()`, `as_GeoPackage()`.
 * Functions to define geoattributes: `define_geoattribute()`, `check_geoattribute_geometry()`, `get_geoattributes()`, `get_geoattribute_geometries()`.
 * Geographic information processing functions: `coordinates_to_point()`, `get_layer_geometry()`, `summarize_layer()`, `get_point_geometry()`.